:recycle: 의존성을 고려한 함수 분기 # 17
